Invocation/OuterPower: track refs by object identity, not clids
This paves the way for synchronous calls, for which an Invocation-side
clist would be really tricky (entries would need to be copied between
the parent Invocation's clist and the child's, on entry and exit from
the child). With a single merged whole-turn refobj->swissnum table, we
can safely hide the swissnums from the inner code, retrieve them from
the (privileged) serializer, and not leak them between frames (unless
they're being explicitly passed in an argument or a return value).