don't let o.call() child grant parent's Memory access to subchildren
The scenario:
 * A does B.call(args.Amem=power.memory), passing A's Memory object into
   B as a regular argument. (B's Power doesn't grant persistent access
   to this Memory, it only gets access because A passed it in on this
   one invocation)
 * B does C=make_urbject(power.memory=args.Amem)
 * later, C is invoked, and modifies it's C.power.memory

The bug was that C.power.memory was the same persistent object as
A.power.memory, giving C persistent access to A's own state. When B
created C, the power.memory= assignment should create a new slot,
populating it with args.Amem as initial contents, but not getting
long-term access to the slot.

This was fixed by making the Turn pay attention to which Invocation is
at the top of the stack, and only comparing the object identity of
power.memory against the one used by the top-most Invocation. Each
o.call() pushes a new Invocation onto the stack, so B isn't allowed to
grant persistent access to a Memory object unless it was endowed via
B.power.memory .